+++ diffly/__init__.py
@@ -11,18 +11,7 @@ except importlib.metadata.PackageNotFoundError as e:  # pragma: no cover
     __version__ = "unknown"
 
 
-from ._exceptions import (
-    CollectionComparisonAssertionError,
-    ComparisonAssertionError,
-    FrameComparisonAssertionError,
-    PrimaryKeyError,
-)
+from ._exceptions import PrimaryKeyError
 from .comparison import compare_frames
 
-__all__ = [
-    "CollectionComparisonAssertionError",
-    "ComparisonAssertionError",
-    "FrameComparisonAssertionError",
-    "PrimaryKeyError",
-    "compare_frames",
-]
+__all__ = ["PrimaryKeyError", "compare_frames"]
